Refactor
- Refactored file structure to use routes files for each route
